+++ rust/host/k16-vm/src/computer/devices.rs
@@ -1,6 +1,6 @@
 mod bios;
 mod control;
-mod framebuffer;
+mod gpu;
 mod keyboard;
 mod serial;
 mod storage;
@@ -9,7 +9,7 @@ mod timer;
 
 pub(crate) use bios::BiosFlashDevice;
 pub(crate) use control::ComputerControlDevice;
-pub(crate) use framebuffer::FramebufferDevice;
+pub(crate) use gpu::GpuDevice;
 pub(crate) use keyboard::validate_event as validate_keyboard_event;
 pub(crate) use keyboard::KeyboardDevice;
 pub use keyboard::KeyboardEvent;